Minor tweaks
ShowdownSet: Lessen allocation
MoveTutor: Remove boxing by calling the generic method instead of object method
Xoro8b: Add more xmldoc, use positive constant instead of inverse negative for parity
StadiumUtil: Use built-in endianness reversal methods

+++ PKHeX.Core/Legality/Moves/MoveTutor.cs
@@ -158,7 +158,7 @@ namespace PKHeX.Core
                 var pi = (PersonalInfoLA)PersonalTable.LA.GetFormEntry(species, form);
                 if (!pi.IsPresentInGame)
                     return NONE;
-                var index = Array.IndexOf(MoveShop8_LA, move);
+                var index = Array.IndexOf(MoveShop8_LA, (ushort)move);
                 if (index != -1 && pi.SpecialTutors[0][index])
                     return GameVersion.PLA;
